aider: Added check for local existence of file before download in Face_Editing.py
# Aider chat conversation:

USER: the download always takes quite some time, can we make a local copy and only download if the local copy does not exist or has changed?

ASSISTANT: To implement this change, we need to modify the `Face_Editing.py` file. We will add a check to see if the file `karras2019stylegan-ffhq-1024x1024.pkl` exists before attempting to download it. If the file does not exist, then we will proceed with the download.

Here is the *SEARCH/REPLACE* block:

Face_Editing.py
```python
<<<<<<< SEARCH
subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'gdown'])
import gdown
url = 'https://drive.google.com/uc?id=1MEGjdvVpUsu1jB4zrXZN7Y4kBBOzizDQ'
output = 'karras2019stylegan-ffhq-1024x1024.pkl'
gdown.download(url, output, quiet=False)
os.system('mv InterFaceGAN/karras2019stylegan-ffhq-1024x1024.pkl InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl')
=======
if not os.path.exists('InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl'):
    subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'gdown'])
    import gdown
    url = 'https://drive.google.com/uc?id=1MEGjdvVpUsu1jB4zrXZN7Y4kBBOzizDQ'
    output = 'karras2019stylegan-ffhq-1024x1024.pkl'
    gdown.download(url, output, quiet=False)
    os.system('mv InterFaceGAN/karras2019stylegan-ffhq-1024x1024.pkl InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl')
else:
    print("File 'karras2019stylegan-ffhq-1024x1024.pkl' already exists. Skipping download.")
>>>>>>> REPLACE
```

+++ Face_Editing.py
@@ -67,12 +67,15 @@ os.chdir('InterFaceGAN/')
 # In[7]:
 
 
-subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'gdown'])
-import gdown
-url = 'https://drive.google.com/uc?id=1MEGjdvVpUsu1jB4zrXZN7Y4kBBOzizDQ'
-output = 'karras2019stylegan-ffhq-1024x1024.pkl'
-gdown.download(url, output, quiet=False)
-os.system('mv InterFaceGAN/karras2019stylegan-ffhq-1024x1024.pkl InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl')
+if not os.path.exists('InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl'):
+    subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'gdown'])
+    import gdown
+    url = 'https://drive.google.com/uc?id=1MEGjdvVpUsu1jB4zrXZN7Y4kBBOzizDQ'
+    output = 'karras2019stylegan-ffhq-1024x1024.pkl'
+    gdown.download(url, output, quiet=False)
+    os.system('mv InterFaceGAN/karras2019stylegan-ffhq-1024x1024.pkl InterFaceGAN/models/pretrain/karras2019stylegan-ffhq-1024x1024.pkl')
+else:
+    print("File 'karras2019stylegan-ffhq-1024x1024.pkl' already exists. Skipping download.")
 
 
 # # I. Let's load our latent space vectors:
